+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Use the window API hook in and out calls from the Chapter.Net.WinAPI assembly.
+### Supported .Net Versions
+- .Net 6 (Windows)
+- .Net 7 (Windows)
+- .Net 8 (Windows)
+
 ## [1.0.0] - 2023-12-23
 ### Added
 - Init project